Renamed io.storage to storage.fs.
